ai_crew/requirements.txt
crewai
langchain-community
requests
python-dotenv
pyyaml
